feat: Add useOutput hook for retrieving output from connected nodes
This commit adds a new hook called useOutput, which retrieves the output from the connected nodes in the application. It takes in a list of connections, nodes, and an options object. The options object allows for customization, with the ability to specify the key to extract the response from the node data. The useOutput hook updates the output whenever there are changes in the connections, nodes, or nodeKey. This hook enhances the functionality of the application by providing a convenient way to access and utilize the output from connected nodes.