feat: Add ItemChestNode component
This commit adds the ItemChestNode component, which renders a Chest node with the ability to add 10 random items to the node's data. The component includes a button to trigger the addition of items and displays the current items in a grid layout. This new component enhances the functionality of the project.