feat: Add support for animated edges
This commit adds support for animated edges by introducing a new edge type called "animated". The code changes include importing the Animated component and adding it to the edgeTypes object. Additionally, the initialEdges array has been updated to use the "animated" type for some edges. This enhancement improves the visual representation of edges in the flow diagram and adds animation functionality to the project.