refactor: Improve debounce function in useStoreExtension
Refactor the debounce function in the useStoreExtension hook to enhance its flexibility and readability. Update the function signature to accept either a function or a string as the first argument. If a string is provided, the corresponding function will be used for debouncing. This change improves the usability of the debounce function and makes it more intuitive for developers to use.